refactor: restructure and rename directories
axios->services, redux->store, enums->constants and move urlConstants to constants
