+++ src/components/NewEntry.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { addEntry } from '../redux/actions';
-import { addMealSuggestionsFor } from '../axios/MealService';
+import { addEntry } from '../store/actions';
+import { addMealSuggestionsFor } from '../services/MealService';
 
 class NewEntry extends React.Component {
   constructor(props) {
