+++ src/tests/components/Entry.test.js
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
-import store from '../../redux/store';
+import store from '../../store/store';
 import Entry from '../../components/Entry';
-import { EntryMode, EntryStatus } from '../../enums/entry';
+import { EntryMode, EntryStatus } from '../../constants/entry';
 
 describe('snapshot tests for Entry component', () => {
   it('should render Entry component with OPEN status and DISPLAY mode correctly', () => {
