+++ src/tests/components/MealSuggestion.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
-import store from '../../redux/store';
+import store from '../../store/store';
 import MealSuggestion from '../../components/MealSuggestion';
 
 describe('snapshot tests for MealSuggestion component', () => {
