+++ src/tests/components/NewEntry.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
-import store from '../../redux/store';
+import store from '../../store/store';
 import NewEntry from '../../components/NewEntry';
 
 describe('snapshot tests for NewEntry component', () => {
